api/requirements.txt
fastapi[standard]
python-dotenv
xrpl-py
httpx
pydantic